chore(backend): Trying to re-organize code
