+++ backend/src/app.module.ts
@@ -2,13 +2,13 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AuthModule } from './auth/auth.module';
+import { AuthModule } from './common/auth/auth.module';
+import { User } from './common/users/user.entity';
+import { UsersModule } from './common/users/users.module';
 import { List } from './lists/list.entity';
 import { ListsModule } from './lists/lists.module';
 import { Task } from './tasks/task.entity';
 import { TasksModule } from './tasks/tasks.module';
-import { User } from './users/user.entity';
-import { UsersModule } from './users/users.module';
 
 @Module({
   imports: [
