+++ backend/src/common/auth/password.module.ts
